+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
+- Methods to work with the activation requests:
+  - `exh.users.activationRequests.find`
+  - `exh.users.activationRequests.remove`
+- Methods to work with the forgot password requests:
+  - `exh.users.forgotPasswordRequests.find`
+  - `exh.users.forgotPasswordRequests.remove`
 - Methods to work with the user verification settings:
   - `exh.users.settings.getVerificationSettings`
   - `exh.users.settings.updateVerificationSettings`